i think someone forgot to run tests. :)

+++ src/main/scala/com/twitter/ostrich/stats/StatsProvider.scala
@@ -188,7 +188,7 @@ trait StatsProvider {
   def timeFutureMicros[T](name: String)(f: Future[T]): Future[T] = {
     val start = Time.now
     f.respond { _ =>
-      addMetric(name + "_usec", start.sinceNow.inMicroseconds)
+      addMetric(name + "_usec", start.sinceNow.inMicroseconds.toInt)
     }
     f
   }
@@ -200,7 +200,7 @@ trait StatsProvider {
   def timeFutureMillis[T](name: String)(f: Future[T]): Future[T] = {
     val start = Time.now
     f.respond { _ =>
-      addMetric(name + "_msec", start.sinceNow.inMilliseconds)
+      addMetric(name + "_msec", start.sinceNow.inMilliseconds.toInt)
     }
     f
   }
@@ -212,7 +212,7 @@ trait StatsProvider {
   def timeFutureNanos[T](name: String)(f: Future[T]): Future[T] = {
     val start = Time.now
     f.respond { _ =>
-      addMetric(name + "_nsec", start.sinceNow.inNanoseconds)
+      addMetric(name + "_nsec", start.sinceNow.inNanoseconds.toInt)
     }
     f
   }
